Validação se employee existe em DeleteEmployeeUseCase

+++ src/modules/employee/useCases/delete-employee.usecase.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { EmployeeRepository } from "../repositories/employee.repository";
 
 @Injectable()
@@ -6,6 +6,12 @@ export class DeleteEmployeeUseCase {
     constructor(private readonly employeeRepository: EmployeeRepository) {}
 
     async execute(id: string) {
+        const employeeExist = await this.employeeRepository.findOne(id);
+
+        if(!employeeExist) {
+            throw new NotFoundException('Employee not found')
+        }
+        
         return await this.employeeRepository.deleteById(id);
     }
 }
